Переносит взаимодействие со store в App
