+++ src/components/map/map.jsx
@@ -2,6 +2,8 @@ import React from "react";
 import PropTypes from "prop-types";
 import leaflet from "leaflet";
 
+const MAP_ZOOM = 12;
+
 class Map extends React.PureComponent {
   constructor(props) {
     super(props);
@@ -14,7 +16,7 @@ class Map extends React.PureComponent {
   componentDidMount() {
     const {city} = this.props;
 
-    const zoom = 12;
+    const zoom = MAP_ZOOM;
 
     this._map = leaflet.map(this._mapRef.current, {
       center: city,
@@ -41,7 +43,7 @@ class Map extends React.PureComponent {
 
   componentDidUpdate() {
     this._markersLayer.clearLayers();
-    this._map.setView(this.props.city, 12);
+    this._map.setView(this.props.city);
     this._renderMarkers();
   }
